tokenssss

+++ ft_atolong.c
@@ -6,15 +6,15 @@
 /*   By: sunakim <sunakim@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/05 21:26:54 by sunakim           #+#    #+#             */
-/*   Updated: 2019/06/05 21:27:07 by sunakim          ###   ########.fr       */
+/*   Updated: 2019/06/05 21:59:22 by sunakim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 long	ft_atolong(char *str)
 {
 	int		i;
-	short	ret;
-	short	neg;
+	long	ret;
+	long	neg;
 
 	i = 0;
 	neg = 1;
